Fixes circumventing deprecated Python and Numpy functionality. Not all modules, such as MMC and KronSVM are currently functional but will be fixed later.

+++ rlscore/learner/_steepest_descent_mmc.pyx
@@ -51,8 +51,7 @@ def findSteepestDirRotateClasses_(double [:, :] Y,
      
     for newclazz in range(labelcount):
          
-        #!!!!!!!!!!!!!!!
-        takenum = (size / labelcount) - classcounts[newclazz] + howmany
+        takenum = int(size / labelcount) - classcounts[newclazz] + howmany
          
         for h in range(takenum):
              
@@ -116,8 +115,7 @@ def findSteepestDirRotateClasses(double [:, :] Y,
     
     for newclazz in range(labelcount):
         
-        #!!!!!!!!!!!!!!!
-        takenum = (size / labelcount) - classcounts[newclazz] + howmany
+        takenum = int(size / labelcount) - classcounts[newclazz] + howmany
         
         for h in range(takenum):
